[Add] 테스트코드 추가 & 테스트 뷰컨 삭제 (#32)
