--- GetARock/GetARock/Screens/Landing/TestViewController.swift
@@ -1,35 +1,0 @@
-//
-//  TestViewController.swift
-//  GetARock
-//
-//  Created by 최동권 on 2023/01/24.
-//
-
-import UIKit
-
-class TestViewController: UIViewController {
-    // 머지 이전 삭제 예정
-    private let collectionView = PositionCollectionView(entryPoint: .position, items: [.position(Position(instrumentName: "기타", imageName: "guitar", isETC: false)),
-                                                                                   .position(Position(instrumentName: "베이스", imageName: "bass", isETC: false)),
-                                                                                   .position(Position(instrumentName: "보컬", imageName: "vocal", isETC: false)),
-                                                                                   .position(Position(instrumentName: "drum", imageName: "drum", isETC: false)),
-                                                                                       .position(Position(instrumentName: "콘트라베이스으으으", imageName: "etc", isETC: false)),
-    ])
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .black
-        view.addSubview(collectionView)
-        collectionView.constraint(top: view.topAnchor,
-                                  leading: view.leadingAnchor, bottom: view.bottomAnchor,
-                                  trailing: view.trailingAnchor,
-        padding: UIEdgeInsets(top: 0, left: 16, bottom: 0, right: 16))
-        collectionView.delegate = self
-    }
-}
-
-extension TestViewController: PositionCollectionViewDelegate {
-   func canSelectPosition(_ collectionView: UICollectionView, indexPath: IndexPath, selectedItemsCount: Int) -> Bool {
-       return true
-   }
-}
